Created Profile Management Page
Created a new page dedicated to Profile Management and moved the Password reset section.
